zeroth order stub of RBG08_rep; +hacks to template
Files for the beginning of this attempt to reproduce RBG08 are
* content.tex
* metadata.yaml
* 191020_circles_GSL_hack.patch (added here for convenience; not used to generate the pdf)
* RBG08_rep.bib
* README.md
* preamble.tex
* article.tex - to include preamble.tex

Since metadata.tex is a generated file, it's removed from the
git repository in this commit; there's also a line added to clean it
in the Makefile, since otherwise changes do not seem to force new
compiles.

A minor hack to article.py, yaml-to-latex.py and rescience.cls
lets the user who has not yet submitted an article get 'DRAFT' in
the top-right corner of each page rather than the misleading
'UNDER REVIEW'. This requires setting a non-empty "drafted:"
line under dates: in the .yaml file. The python Date parsing
used here seems to be locale-sensitive (for me it considers any
reasonable English or decimal date to be empty), so this hack does not
attempt to parse the dates["drafted"] entry as a date.