Update text to match steps done
Describe fortran stricter standard fix; GSL downgrade.

Need for a script as part of the method and also a result
in terms of how even my notes were not complete enough to
reproduce the results trivially.

References added, including a subtle hint for why we should
gradually shift off git repository servers that violate the ISC's
principles of the universal right to access to science.