Add pdf with meaningful name
Git repositories should not normally be used for distributing
produced files, but ReScience C requires a pdf to be provided,
and since the pdf of this paper is only abou 80 kb in size,
this does not add much server load.

Thus, this commit adds the produced pdf `Roukema_ReSciC2020.pdf`
to the repository, along with a trivial Makefile modification.
You can safely delete the pdf and recreate it in your working
tree.