Free-licensing the ReScience C template
Until this commit, although the fonts in the font subdirectories
are clearly free-licensed, the python scripts, tex templates,
README.md, and other auxiliary files are not free-licensed.  In
principle, authors modify the template to prepare papers for
submission to ReScience C have... violated Nicolas Rougier's
copyright on the files, since the conditions for modifying and
redistributing the files are "whatever Nicolas [and others who
have edited the files] would consider reasonable", with no
guarantee of the right to redistribute the original or modified
files arbitrarily - or sell them.

This commit proposes a modification to the README.md, adding a
GPL-3+ copyright for the files that are not already covered by a
free licence. This would make it easier for people to prepare
papers for ReScience C and to post their drafts on git servers
that only allow free licensed code to be deposited, such as
https://codeberg.org .