Fix Makefile bug; +swh; Reproduction section.
The default ReScience C Makefile cannot handle URLs
with '%' escaped special characters, such as %26 for
the ampersand. A small hack in the Makefile fixes this,
which is necessary for Astronomy & Astrophysics URLs.

The Software Heritage (swh) URL of the code at codeberg
is added to the metadata:
https://archive.softwareheritage.org/browse/origin/https://codeberg.org/boud/0807.4260/

ReScience C now has a fourth section - Reproduction - which seems
appropriate for the Ten Years Reproducibility Challenge.