Discuss AC_F77_WRAPPERS and hunt for g77 source; +substructure
With this commit, the key elements of the comments under the
submitted version have been integrated:

* the autoconf recommendations on using AC_F77_WRAPPERS for
Fortran 77 (g77) -- C (gcc) interfacing are discussed with
references; the Oxford Uni Physics Dept opinion that this
interfacing was difficult is added;

* the apparent absence of archived source code for g77 is
explained, together with an example of where the source can
be found.

Section 3.4 on running is renamed and split into subsubsections
for more clarity.