Revised version in response to reviewer comments
Several references have been added to clarify that the git
repository bans on users in some countries and territories are
not just hypothetical statements made as legal formalities, they
are real bans that have been enacted and obstructed scientific
freedom; and the ethical problem is raising concern in the
community.

A comment on the possibility of bypassing the fortran 77 - C
interface issues by using f2c has been added.

The title has had the [[not-symbol=xac]Rp] tag added.