Qualify partial block; fix funny author name
This commit adds the word `partially` in the Method section,
paragraph 2, to clarify that the Github/Bitbucket/Gitlab bans on
selected countries/territories do not appear to be full bans.
In the absence of known research results documenting
the details of the effectiveness and completeness of the bans,
there is little point speculating on this.

The `I. S. Council` as an author has been fixed to read
`International Science Council`.