Proofread: missing DOIs; hyphens/en dashes
Proofread: Journals normally have a proofreading stage
after acceptance, prior to the final version of record.

The use of DOIs to generate URLs is nice, but proofreaders would
in that case normally require the author to add missing DOIs to
the .bib file (or produced file), since otherwise the articles
that are formally published research articles and have DOIs would
look odd. This commit adds the missing DOIs.

A proofreader would also have noticed the error with hyphens and
en dashes in the list of libraries. This is corrected in this
commit.

Since it is still 14 July 2020 in most of the world, this
commit will hopefully be counted in the version of record
by the Editor. :)

*  modified:   RBG08_rep.bib
*  modified:   Roukema_ReSciC2020.pdf
*  modified:   content.tex