DOI: oadoi.org; ArXiv identifiers
Proofreading changes:

This commit changes the DOI links in the references to open
access DOI oadoi.org for the link and displays the DOI identifier
in the text. Related changes: \sloppy option for references is
inserted, to avoid having text extending into the margin;
(most) URL texts are made copy/pastable by redefining \nolinkurl
to print text without alteration.

A better version of this last fix should probably found before
using this hack for other articles.  (Examples where the text URL
is unusable are refs [12] with
`https : / / techcrunch . com / 2019 / 07 / 29 / github - ban -`
and [24] with
`https : / / gcc . gnu . org /`.

The text appearance of the eprint identifers was wrong because
it inserted a space into the identifier; see
https://arxiv.org/help/arxiv_identifier . This is fixed in
this commit.

The display of URLs is enabled. Linking to the URL archives
is good, but the original URLs should be shown too - archives
can themselves be sometimes unavailable, not only URLs.

* modified:   Roukema_ReSciC2020.pdf
* modified:   rescience.cls