Ordinary LaTeX version
Reproducibility includes the idea of *not* assuming that the user
has updated to the most recent tools. This commit produces an
ordinary latex file, so that automatic pipelines such as the
ArXiv pipeline only need the .tex file (in this case, a produced
file called 'Roukema_ReSciC2020.tex'), rescience.cls, and
metadata.tex (which is read in by rescience.cls).  These are
archived in a file 'arXiv.tgz'. Part of the Makefile rule is
hardwired to handle url breaking that is not handled correctly
by breakurl and hyperref.